# backend/docs/swagger.yaml
basePath: /api
definitions:
...
  entities.Availability:
    properties:
      chair:
        allOf:
        - $ref: '#/definitions/entities.Chair'
        description: Relacionamentos
      chair_id:
        type: integer
      created_at:
        type: string
      day_of_week:
        description: 0=Domingo, 6=Sábado
        maximum: 6
        minimum: 0
        type: integer
      end_time:
        description: Formato HH:MM
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      start_time:
        description: Formato HH:MM
        type: string
      updated_at:
        type: string
      valid_from:
        type: string
      valid_to:
        type: string
    required:
    - chair_id
    - day_of_week
    - end_time
    - start_time
    type: object
  entities.Booking:
    properties:
      chair:
        $ref: '#/definitions/entities.Chair'
      chair_id:
        type: integer
      created_at:
        type: string
      end_time:
        type: string
      id:
        type: integer
      notes:
        type: string
      start_time:
        type: string
      status:
        enum:
        - agendado
        - confirmado
        - cancelado
        - concluido
        - falta
        type: string
      updated_at:
        type: string
      user:
        allOf:
        - $ref: '#/definitions/entities.User'
        description: Relacionamentos
      user_id:
        type: integer
    required:
    - chair_id
    - end_time
    - start_time
    - user_id
    type: object
  entities.Chair:
    properties:
      availabilities:
        items:
          $ref: '#/definitions/entities.Availability'
        type: array
      bookings:
        description: Relacionamentos
        items:
          $ref: '#/definitions/entities.Booking'
        type: array
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      location:
        maxLength: 100
        minLength: 2
        type: string
      name:
        maxLength: 100
        minLength: 2
        type: string
      status:
        enum:
        - ativa
        - inativa
        type: string
      updated_at:
        type: string
    required:
    - location
    - name
    type: object
  entities.User:
    properties:
      birth_date:
        type: string
      bookings:
        description: Relacionamentos
        items:
          $ref: '#/definitions/entities.Booking'
        type: array
      cpf:
        type: string
      created_at:
        type: string
      email:
        type: string
      function:
        type: string
      gender:
        enum:
        - masculino
        - feminino
        - outro
        type: string
      id:
        type: integer
      last_login:
        type: string
      name:
        maxLength: 60
        minLength: 2
        type: string
      phone:
        maxLength: 20
        minLength: 10
        type: string
      position:
        type: string
      registration:
        type: string
      requested_role:
        enum:
        - usuario
        - atendente
        - admin
        type: string
      role:
        enum:
        - usuario
        - atendente
        - admin
        type: string
      sector:
        type: string
      status:
        enum:
        - pendente
        - aprovado
        - reprovado
        type: string
      updated_at:
        type: string
    required:
    - cpf
    - email
    - name
    - phone
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: API REST para sistema de agendamento de cadeiras de massagem
  title: Sistema de Agendamento API
  version: "1.0"
paths:
...
  /availabilities:
    get:
      consumes:
      - application/json
      description: Lista disponibilidades com paginação e filtros opcionais
      parameters:
      - default: 10
        description: Limite de registros por página
        in: query
        name: limit
        type: integer
      - default: 0
        description: Offset para paginação
        in: query
        name: offset
        type: integer
      - description: Filtrar por ID da cadeira
        in: query
        name: chair_id
        type: integer
      - description: Filtrar por dia da semana (0-6)
        in: query
        name: day_of_week
        type: integer
      - description: Filtrar por status ativo
        in: query
        name: is_active
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: Lista de disponibilidades
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Parâmetros inválidos
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Token inválido
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Listar disponibilidades
      tags:
      - availabilities
    post:
      consumes:
      - application/json
      description: Cria uma nova disponibilidade para uma cadeira (apenas admins)
      parameters:
      - description: Dados da disponibilidade
        in: body
        name: availability
        required: true
        schema:
          $ref: '#/definitions/entities.Availability'
      produces:
      - application/json
      responses:
        "201":
          description: Disponibilidade criada com sucesso
          schema:
            $ref: '#/definitions/entities.Availability'
        "400":
          description: Dados inválidos
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Token inválido
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Sem permissão
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Criar disponibilidade
      tags:
      - availabilities
  /availabilities/{id}:
    get:
      consumes:
      - application/json
      description: Retorna os dados de uma disponibilidade específica pelo ID
      parameters:
      - description: ID da disponibilidade
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Disponibilidade encontrada
          schema:
            $ref: '#/definitions/entities.Availability'
        "400":
          description: ID inválido
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Token inválido
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Disponibilidade não encontrada
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Buscar disponibilidade por ID
      tags:
      - availabilities
    put:
      consumes:
      - application/json
      description: Atualiza os dados de uma disponibilidade existente (apenas admins)
      parameters:
      - description: ID da disponibilidade
        in: path
        name: id
        required: true
        type: integer
      - description: Dados atualizados da disponibilidade
        in: body
        name: availability
        required: true
        schema:
          $ref: '#/definitions/entities.Availability'
      produces:
      - application/json
      responses:
        "200":
          description: Disponibilidade atualizada com sucesso
          schema:
            $ref: '#/definitions/entities.Availability'
        "400":
          description: Dados inválidos
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Token inválido
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Sem permissão
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Disponibilidade não encontrada
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Atualizar disponibilidade
      tags:
      - availabilities
...
  /chairs/{chair_id}/availabilities:
    get:
      consumes:
      - application/json
      description: Retorna todas as disponibilidades configuradas para uma cadeira
        específica
      parameters:
      - description: ID da cadeira
        in: path
        name: chair_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Disponibilidades da cadeira
          schema:
            items:
              $ref: '#/definitions/entities.Availability'
            type: array
        "400":
          description: ID da cadeira inválido
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Token inválido
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Cadeira não encontrada
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Buscar disponibilidades de uma cadeira
      tags:
      - availabilities
  /chairs/{chair_id}/available-slots:
    get:
      consumes:
      - application/json
      description: Retorna os horários disponíveis para agendamento em uma cadeira
        específica em uma data
      parameters:
      - description: ID da cadeira
        in: path
        name: chair_id
        required: true
        type: integer
      - description: 'Data para buscar horários (formato: YYYY-MM-DD)'
        in: query
        name: date
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Lista de horários disponíveis
          schema:
            items:
              type: string
            type: array
        "400":
          description: Parâmetros inválidos
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Token inválido
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Buscar horários disponíveis
      tags:
      - availabilities
...
securityDefinitions:
  Bearer:
    description: Digite 'Bearer' seguido de um espaço e o token JWT obtido no login
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"